+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Removed the @Resource hook. The behaviour was a bit iffy and the same functionality can mostly be achieved using `toJSON()`
 * Renamed the `basePath` service config option to `prefix`
 * Routes are now defined within the `routes` member variable of service classes
+* `index()` function of service classes gets called for `GET /` instead of `get()` (which now only gets called for `GET/:id`)
 
 ## 1.0.0 (21st March 2021)
 * Bump to 1.0.0 (the project feels feature-complete now)